requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20230802.0=h313beb8_2
aiohttp=3.9.0=py310h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=4.2.0=pyhd8ed1ab_0
arrow-cpp=11.0.0=hc7aafb3_2
async-timeout=4.0.3=py310hca03da5_0
attrs=23.1.0=py310hca03da5_0
aws-c-common=0.6.8=h80987f9_1
aws-c-event-stream=0.1.6=h313beb8_6
aws-checksums=0.1.11=h80987f9_2
aws-sdk-cpp=1.8.185=ha71a6ea_1
blas=1.0=openblas
boost-cpp=1.82.0=h48ca7d4_2
bottleneck=1.3.5=py310h96f19d2_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
bzip2=1.0.8=h93a5062_5
c-ares=1.19.1=h80987f9_0
ca-certificates=2023.12.12=hca03da5_0
catalogue=2.0.10=py310hbe9552e_0
certifi=2023.11.17=py310hca03da5_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=py310hca03da5_0
cloudpathlib=0.16.0=pyhd8ed1ab_0
colorama=0.4.6=py310hca03da5_0
confection=0.1.4=py310ha73103e_0
cymem=2.0.8=py310h1253130_1
cython-blis=0.7.10=py310h280b8fa_2
dataclasses=0.8=pyhc8e2a94_3
datasets=2.12.0=py310hca03da5_0
dill=0.3.6=py310hca03da5_0
distro=1.8.0=py310hca03da5_0
en-core-web-sm=3.7.1=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
fastapi=0.99.1=pyhd8ed1ab_0
filelock=3.13.1=py310hca03da5_0
freetype=2.12.1=h1192e45_0
frozenlist=1.4.0=py310h80987f9_0
fsspec=2023.10.0=py310hca03da5_0
gensim=4.3.2=py310h5924a0a_0
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_3
glog=0.5.0=hc377ac9_0
gmp=6.3.0=h965bd2d_0
gmpy2=2.1.2=py310h8c48613_0
grpc-cpp=1.48.2=hc60591f_4
gtest=1.14.0=h48ca7d4_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=py_0
httpcore=1.0.2=pyhd8ed1ab_0
httpx=0.26.0=pyhd8ed1ab_0
huggingface_hub=0.17.3=py310hca03da5_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=73.1=h313beb8_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.0=py310hca03da5_1
isort=5.13.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=py310hca03da5_0
jpeg=9e=h80987f9_1
krb5=1.20.1=hf3e1bf2_1
langcodes=3.3.0=pyhd8ed1ab_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libboost=1.82.0=h0bc93f9_2
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcurl=8.4.0=h3e2b118_1
libcxx=16.0.6=h4653b0c_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=13_2_0_hd922786_1
libgfortran5=13.2.0=hf226fd6_1
libiconv=1.16=h1a28f6b_2
libnghttp2=1.57.0=h62f6fdd_0
libopenblas=0.3.25=openmp_h6c19121_0
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libsqlite=3.44.2=h091b4b1_0
libssh2=1.10.0=h02f6b3c_2
libthrift=0.15.0=h73c2103_2
libtiff=4.5.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=17.0.6=hcd81f8e_0
lz4-c=1.9.4=hb7217d7_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.3=py310h80987f9_0
mdurl=0.1.2=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h9546428_0
mpmath=1.3.0=py310hca03da5_0
multidict=6.0.4=py310h80987f9_0
multiprocess=0.70.14=py310hca03da5_0
murmurhash=1.0.10=py310h1253130_1
ncurses=6.4=h463b476_2
networkx=3.2.1=pyhd8ed1ab_0
nltk=3.8.1=pyhd8ed1ab_0
numexpr=2.8.7=py310hecc3335_0
numpy=1.26.3=py310h3b2db8e_0
numpy-base=1.26.3=py310ha9811e2_0
openai=1.6.1=pyhd8ed1ab_0
openjpeg=2.3.0=h7a6adac_2
openssl=3.2.0=h0d3ecfb_1
orc=1.7.4=hdca1487_1
packaging=23.1=py310hca03da5_0
pandas=2.1.4=py310h46d7db6_0
pathy=0.10.2=pyhd8ed1ab_0
pillow=10.0.1=py310h3b245a6_0
pip=23.3.2=pyhd8ed1ab_0
preshed=3.0.9=py310h1253130_1
pyarrow=11.0.0=py310hbfed03b_1
pydantic=1.10.12=py310h80987f9_1
pygments=2.17.2=pyhd8ed1ab_0
python=3.10.13=h2469fbe_1_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-multipart=0.0.6=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py310h1a28f6b_1
python_abi=3.10=4_cp310
pytorch=2.1.2=py3.10_0
pytz=2023.3.post1=py310hca03da5_0
pyyaml=6.0.1=py310h80987f9_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h92ec313_1
regex=2023.10.3=py310h80987f9_0
requests=2.31.0=pyhd8ed1ab_0
responses=0.13.3=pyhd3eb1b0_0
rich=13.7.0=pyhd8ed1ab_0
safetensors=0.4.0=py310h482802a_0
scikit-learn=1.3.0=py310h46d7db6_0
scipy=1.11.4=py310h20cbe94_0
sentence-transformers=2.2.2=pyhd8ed1ab_0
sentencepiece=0.1.99=py310h48ca7d4_0
setuptools=68.2.2=py310hca03da5_0
shellingham=1.5.4=pyhd8ed1ab_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py310hca03da5_0
snappy=1.1.10=h313beb8_1
sniffio=1.3.0=pyhd8ed1ab_0
spacy=3.7.2=py310h2a8a171_0
spacy-legacy=3.0.12=pyhd8ed1ab_0
spacy-loggers=1.0.5=pyhd8ed1ab_0
sqlite=3.41.2=h80987f9_0
srsly=2.4.8=py310h1253130_1
starlette=0.27.0=py310hca03da5_0
sympy=1.12=py310hca03da5_0
thinc=8.2.2=py310h1359cc7_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.13=h5083fa2_1
tokenizers=0.13.2=py310h3dd52b7_1
torchaudio=2.1.2=py310_cpu
torchvision=0.15.2=cpu_py310h31aa045_0
tqdm=4.66.1=pyhd8ed1ab_0
transformers=4.32.1=py310hca03da5_0
typer=0.9.0=pyhd8ed1ab_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023d=h0c530f3_0
urllib3=2.1.0=py310hca03da5_0
utf8proc=2.6.1=h1a28f6b_0
uvicorn=0.20.0=py310hca03da5_0
wasabi=1.1.2=py310hbe9552e_0
weasel=0.3.4=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
xxhash=0.8.0=h1a28f6b_3
xz=5.4.5=h80987f9_0
yaml=0.2.5=h3422bc3_2
yarl=1.9.3=py310h80987f9_0
zipp=3.17.0=py310hca03da5_0
zlib=1.2.13=h53f4e23_5
zss=1.2.0=pypi_0
zstd=1.5.5=hd90d995_0
